requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
httpx[http2]==0.27.2
tenacity==9.0.0
beautifulsoup4==4.12.3


